Refactor game firebase functions
